@@ -14,43 +14,43 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.3.11",
-    "@angular/cdk": "~13.3.1",
-    "@angular/common": "13.3.11",
-    "@angular/compiler": "13.3.11",
-    "@angular/core": "13.3.11",
-    "@angular/flex-layout": "13.0.0-beta.38",
-    "@angular/forms": "13.3.11",
-    "@angular/material": "~13.3.1",
-    "@angular/platform-browser": "13.3.11",
-    "@angular/platform-browser-dynamic": "13.3.11",
-    "@angular/router": "13.3.11",
-    "@auth0/angular-jwt": "5.0.2",
-    "@ngrx/store": "13.1.0",
-    "ngx-device-detector": "^3.0.0",
-    "rxjs": "7.5.5",
-    "tslib": "2.3.1",
-    "zone.js": "0.11.5"
+    "@angular/animations": "17.0.6",
+    "@angular/cdk": "~17.0.3",
+    "@angular/common": "17.0.6",
+    "@angular/compiler": "17.0.6",
+    "@angular/core": "17.0.6",
+    "@angular/flex-layout": "15.0.0-beta.42",
+    "@angular/forms": "17.0.6",
+    "@angular/material": "~17.0.3",
+    "@angular/platform-browser": "17.0.6",
+    "@angular/platform-browser-dynamic": "17.0.6",
+    "@angular/router": "17.0.6",
+    "@auth0/angular-jwt": "5.2.0",
+    "@ngrx/store": "17.0.1",
+    "ngx-device-detector": "^7.0.0",
+    "rxjs": "7.8.1",
+    "tslib": "2.6.2",
+    "zone.js": "0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "13.3.2",
-    "@angular/cli": "13.3.2",
-    "@angular/compiler-cli": "13.3.11",
-    "@angular/language-service": "13.3.11",
-    "@types/jasmine": "4.0.0",
-    "@types/jasminewd2": "2.0.10",
+    "@angular-devkit/build-angular": "17.0.6",
+    "@angular/cli": "17.0.6",
+    "@angular/compiler-cli": "17.0.6",
+    "@angular/language-service": "17.0.6",
+    "@types/jasmine": "5.1.4",
+    "@types/jasminewd2": "2.0.13",
     "codelyzer": "6.0.2",
-    "jasmine-core": "4.0.1",
+    "jasmine-core": "5.1.1",
     "jasmine-spec-reporter": "7.0.0",
-    "karma": "6.3.17",
-    "karma-chrome-launcher": "3.1.1",
+    "karma": "6.4.2",
+    "karma-chrome-launcher": "3.2.0",
     "karma-coverage-istanbul-reporter": "3.0.3",
-    "karma-jasmine": "4.0.1",
-    "karma-jasmine-html-reporter": "1.7.0",
-    "ng-openapi-gen": "0.21.2",
+    "karma-jasmine": "5.1.0",
+    "karma-jasmine-html-reporter": "2.1.0",
+    "ng-openapi-gen": "0.51.0",
     "protractor": "7.0.0",
-    "ts-node": "10.7.0",
+    "ts-node": "10.9.2",
     "tslint": "5.20.1",
-    "typescript": "~4.6.3"
+    "typescript": "~5.2.0"
   }
 }
