@@ -16,62 +16,63 @@
     "test:debug": "node --inspect -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand"
   },
   "dependencies": {
-    "@nestjs/axios": "^0.1.0",
-    "@nestjs/common": "8.4.4",
-    "@nestjs/core": "8.4.4",
-    "@nestjs/jwt": "8.0.0",
-    "@nestjs/microservices": "8.4.4",
-    "@nestjs/passport": "8.2.1",
-    "@nestjs/platform-express": "8.4.4",
-    "@nestjs/swagger": "5.2.1",
-    "@nestjs/websockets": "8.4.4",
-    "@types/http-proxy": "1.17.8",
+    "@nestjs/axios": "^3.0.1",
+    "@nestjs/common": "10.2.10",
+    "@nestjs/core": "10.2.10",
+    "@nestjs/jwt": "10.2.0",
+    "@nestjs/microservices": "10.2.10",
+    "@nestjs/passport": "10.0.3",
+    "@nestjs/platform-express": "10.2.10",
+    "@nestjs/swagger": "7.1.16",
+    "@nestjs/websockets": "10.2.10",
+    "@types/http-proxy": "1.17.14",
     "@types/http-proxy-middleware": "0.19.3",
-    "@types/nodemailer": "6.4.4",
-    "@types/tar": "^6.1.3",
+    "@types/nodemailer": "6.4.14",
+    "@types/tar": "^6.1.10",
     "epub": "1.2.1",
-    "flat": "^5.0.2",
-    "http-proxy-middleware": "^2.0.4",
+    "flat": "^5.0.0",
+    "http-proxy-middleware": "^2.0.6",
     "jsonfile": "^6.1.0",
     "libgen": "^2.1.2",
-    "memfs": "3.4.1",
+    "memfs": "^4.6.0",
     "morgan": "1.10.0",
     "ncp": "2.0.0",
     "node-calibre": "2.1.1",
-    "nodemailer": "6.7.3",
-    "passport": "0.5.2",
-    "passport-jwt": "4.0.0",
+    "nodemailer": "6.9.7",
+    "passport": "0.7.0",
+    "passport-jwt": "4.0.1",
     "passport-local": "1.0.0",
-    "reflect-metadata": "0.1.13",
-    "rxjs": "7.5.5",
-    "swagger-ui-express": "4.3.0",
-    "tar": "^6.1.12"
+    "reflect-metadata": "0.1.14",
+    "rxjs": "7.8.1",
+    "swagger-ui-express": "5.0.0",
+    "tar": "^6.2.0"
   },
   "devDependencies": {
-    "@nestjs/cli": "^8.2.5",
-    "@nestjs/testing": "8.4.4",
-    "@types/express": "4.17.13",
-    "@types/flat": "^5.0.2",
-    "@types/jest": "27.4.1",
-    "@types/jsonfile": "^6.1.0",
-    "@types/morgan": "1.9.3",
-    "@types/multer": "1.4.7",
-    "@types/node": "17.0.23",
-    "@types/passport-jwt": "3.0.6",
-    "@types/passport-local": "1.0.34",
-    "@types/supertest": "2.0.12",
-    "jest": "28.1.0",
-    "mkdirp": "1.0.4",
-    "prettier": "2.8.8",
-    "rimraf": "3.0.2",
-    "supertest": "6.2.2",
-    "ts-jest": "28.0.2",
-    "ts-loader": "9.2.8",
-    "ts-node": "10.7.0",
-    "tsconfig-paths": "3.14.1",
+    "@nestjs/cli": "^10.2.1",
+    "@nestjs/testing": "10.2.10",
+    "@types/express": "4.17.21",
+    "@types/jest": "29.5.11",
+    "@types/flat": "^5.0.5",
+    "@types/jsonfile": "^6.1.4",
+    "@types/morgan": "1.9.9",
+    "@types/multer": "1.4.11",
+    "@types/node": "^18",
+    "@types/passport-jwt": "3.0.13",
+    "@types/passport-local": "1.0.38",
+    "@types/supertest": "2.0.16",
+    "@tsconfig/node18": "18.2.2",
+    "jest": "29.7.0",
+    "mkdirp": "3.0.1",
+    "prettier": "3.1.0",
+    "rimraf": "5.0.5",
+    "supertest": "6.3.3",
+    "ts-jest": "29.1.1",
+    "ts-loader": "9.5.1",
+    "ts-node": "^10.9.1",
+    "tsconfig-paths": "4.2.0",
     "tslint": "5.20.1",
-    "typescript": "4.6.3",
-    "webpack": "^5.70.0"
+    "typescript": "5.3.3",
+    "webpack": "^5.89.0"
   },
   "jest": {
     "moduleFileExtensions": [
