@@ -3,8 +3,8 @@ import path from 'path';
 import { ConfigMissingException } from '../lib/exceptions';
 import Config from './config.default.json';
 import { randomBytes } from 'crypto';
-import jsonfile from 'jsonfile';
-import flat from 'flat';
+import { writeFileSync, readFileSync } from 'jsonfile';
+import { flatten, unflatten } from 'flat';
 
 @Injectable()
 export class ConfigService {
@@ -13,20 +13,20 @@ export class ConfigService {
     constructor(
         private readonly log?: Logger,
     ) {
-        let configFlat = this.replaceValuesWithWildcard(flat.flatten<any, any>(Config, { delimiter: this.DELIMITER }), __dirname);
-        this.config = flat.unflatten<any, any>(configFlat, { delimiter: this.DELIMITER });
+        let configFlat = this.replaceValuesWithWildcard(flatten<any, any>(Config, { delimiter: this.DELIMITER }), __dirname);
+        this.config = unflatten<any, any>(configFlat, { delimiter: this.DELIMITER });
         const envConfigFlat = ConfigService.mapEnvToObject();
-        const envConfig = flat.unflatten<any, any>(envConfigFlat, { delimiter: this.DELIMITER });
+        const envConfig = unflatten<any, any>(envConfigFlat, { delimiter: this.DELIMITER });
         this.config.app.configOverridePath = envConfig.app?.configOverridePath || this.config.app.configOverridePath;
 
         this.loadConfig(this.config.app.configOverridePath);
-        configFlat = flat.flatten<any, any>(this.config, { delimiter: this.DELIMITER });
+        configFlat = flatten<any, any>(this.config, { delimiter: this.DELIMITER });
         const resultFlat = {
             ...configFlat,
             ...envConfigFlat,
         };
 
-        this.config = flat.unflatten<any, any>(resultFlat, { delimiter: this.DELIMITER });
+        this.config = unflatten<any, any>(resultFlat, { delimiter: this.DELIMITER });
         this.config.auth.jwtSecret = randomBytes(20).toString();
         this.log?.log(`Succesfully loaded config: ${JSON.stringify(this.config)}`);
     } 
@@ -35,26 +35,26 @@ export class ConfigService {
     public loadConfig(filePath: string): boolean {
         let configOverrideFlat = {};
         try {
-            configOverrideFlat = flat.flatten<any, any>(jsonfile.readFileSync(filePath), { delimiter: this.DELIMITER });
+            configOverrideFlat = flatten<any, any>(readFileSync(filePath), { delimiter: this.DELIMITER });
         } catch (e) {
             this.log?.warn(`Could not read config file ${filePath} does not exist, using defaults`);
             return false;
         }
         
-        const curConfigFlat = flat.flatten<any, any>(this.config, { delimiter: this.DELIMITER });
+        const curConfigFlat = flatten<any, any>(this.config, { delimiter: this.DELIMITER });
         let resultFlat = {
             ...curConfigFlat,
             ...configOverrideFlat,
         }
         resultFlat = this.replaceValuesWithWildcard(resultFlat, path.resolve(path.dirname(filePath)));
 
-        this.config = flat.unflatten<any, any>(resultFlat, { delimiter: this.DELIMITER });
+        this.config = unflatten<any, any>(resultFlat, { delimiter: this.DELIMITER });
         this.log?.log(`Successfully loaded config from file: ${filePath}`);
         return true;
     }
 
     public saveConfig(): void {
-        jsonfile.writeFileSync(
+        writeFileSync(
             this.config.app.configOverridePath,
             this.config,
             {
